refactor(cli): Phase 3 follow-ups for unified --format option
Bounded refactor across 6 items for Phase 3 --format option cleanup:

1. Consolidate yaml imports: Replace conditional imports with lazy _get_yaml()
   helper in flow_status.py, handoff_read.py, and task_ui.py
2. Rename format→output_format: Rename function parameters to avoid shadowing
   Python builtin, update all internal references
3. Fix aborted flow json/yaml: Add structured output for aborted flows
   (previously fell through to timeline path)
4. Apply filtering to structured output: Apply --show-all and --actor filtering
   to json/yaml events (previously unfiltered)
5. Defer JsonOption cleanup: Record in handoff for separate Phase 3 effort
6. Stabilise test parsing: Use result.stdout/stderr instead of splitting
   result.output (Click provides separate attributes by default)

Additional refactoring:
- Extract helper functions from flow_status.py to flow_status_helpers.py
- Reduce flow_status.py from 434 to 278 lines (under 400-line CI threshold)

All tests pass: pytest=PASS (22/22), mypy=PASS, ruff=PASS

Closes #751
